Renamed LibFS project LibCommon, will be used to store common types between bootloader and kernel.
Added Checks to stop redefinition of int123 and uint128 struct in any projects that use both UEFI and LibCommon libs together.

Goal for LibCommon is to remove need for UEFI from it and have the bootloader build the needed objects to start kernel, the kernel will have no knowledge of the uefi other than pointers to some functions that are retained after BootServices.Exit() is called.

TODO: New C++ Project Called LibRT which will contain common functions/types and other code for use in the kernel and the core OS services/drivers.