Switched DrawRectangle and DrawFilledRectangle as i dont have an outlined rectangle function at the moment.
Added Thickness parameter to Non-Filled Draw Functions to specifiy the outlines thickness.

Created a Font.h as a marker for where to start a basic font system for loading a basic set of fonts for the Kernel and bootloader to use in Graphics mode.

TODO: Look into Serial output/input and Debug protocols to see if i can get step over, break and stepped debugging running so i can more easily diagnose issues.