Renamed EFI_FS.h to FileSystemContext.h
Added New Functions for getting File and Volume Info from the FileSystemContext.

Added functions to FileInfo and Volume Info and have hidden the constructors  as protected so that they are behind a function and can't be created by anything but the library code.

Added some code to bootloader to search for the "Sys" partition/volume on the drive and mark as found/not found as a test.

TODO: Start Process of loading a file(kernel) into memory and properly jumping into execution(Call ExitBootServices after setting up the system for the kernel to take over and start the process of loading services & drivers)

Write a new Library for Services that are currently relying on EFI Calls(GOP,FileSystem,etc) for the kernel.(Most likey LibKrnl will be the name) this libray will handle alot of c++ runtime stuff that is unusable to the bootloader at the moment (malloc, new,delete, memcpy,memcmp,etc.), but will be needed when we transition from bootloader to kernel.