Refactored PE32 File Type support.
Moed all COFF/MSDOS/PE32/PE32+ Header Information into single File called PE32.h and source under PE32.cpp, the provided headers will allow reading of DOS & PE32/PE32+ Compatible Files Into memory using the functions found under the PE32File Class, this takes advantage of the the allocator overloads provided in the Updated Allocator.h file. there is currently a but with calling delete on structs, that will cause a GPF, this is being investigated as it's most likely an error in the overloads for the delete functions, this will be sorted for the next commit.

Next Goal is to implement PESectionHeader Reading Support to allow for proper loading of the binaries into memory for execution.