Finished writing Program and section headers for ELF binaires, add constructors for loading data from a ESP File Handle.
Added some new functions to FileHandle as templates to make it easier to write code that loads in data of KnownSized structs and arrays of those structs.

rewritten most of the file reading code to use these new templates to make it more clear as to what is loading.

Started process to comment on every structure and function in the LibCommon Library.

Next goal is to get a font FileType added so thta we can start the process of loading a font at boot.