Move PSF into PCSF folder, Added version 2 headers and finished writing the code to read the headers in, need to get some test fonts to make sure this code works correctly and to adjust how im going to read in the unicode tables in a consistent and friendly way.
Qorking on making the Qemu manager become able to read a json file for settings and allowing it to optionally automatically pull the latest OVMF binairies from a git repo for the testing, or just a set version. eventually want to make the tool able to do most of the work of setting up the dev enviroment and potentually make a better GUI for running it.

may also make a couple othewr tools for easier debugging from Visual Studios through QEMU, but that is probably going to be a seperate repo.