Updated project for enhanced graphics and debugging capabilities
This commit includes a major overhaul of the graphics system, replacing `GraphicsContext` with `RenderContext` and `MonitorContext` in various files. New pixel structures were added to represent different bit depths. The `DrawLine` and `DrawRectangle` functions were updated, and a new `PrintWarning` function was added. The `Math.h` library was replaced with `Numerics/Math.h` and the namespace `Common::Numerics` was added.  The `MonitorMode` constructor was modified to accept a `EFI_GRAPHICS_OUTPUT_PROTOCOL_MODE*` parameter. Several changes were made to `MonitorContext.cpp` and `MonitorContext.h`, including renaming `currentMode` to `currentModeNumber`, adding a new `currentMode` pointer, and adding new functions. The `Colour` struct was updated to use `#pragma pack` directives to control the alignment of its members. The `Math.h` file was replaced with a new implementation of the `Math` class in the `Common::Numerics` namespace. A new interface `IConsole` was added in `IConsole.h`. The definitions for `UINTN` and `INTN` were changed in `TypeDefs.h` and `UEFIDef.h`. In `Program.cs`, the `UpdateOVMF` method was simplified and the `qemu` process now waits for exit and checks the exit code.

New Vector class added for later use for the system when dealing with certain types of data.

Qemu Manager updates have been changed to only use the commit sha as it simplies the checks.

KNOWN ISSUE:
Currently for some reason, even though the kernel is loaded and the render does not throw any messages, it appears to not be drawing to the screen or clearing the framebuffer. not usre what is causing this issue at this time and have tried a few diffrent things to attempt to fix it. the GraphicsContext has been kept for now as a backup incase i cannot get the RenderContext and MonitorContext Systems to behave.