Moved some of the code related to the File System around as most of it can be reused, renamed the FileSystemContext to ESP_FS_Context to reflect the filesystem that it is written around.
made changes in some files to reflect the changes done for the FileSystem

Moved the Font File Definitions into Graphics folder.

Reworked ther PCSF1 Font file format so that it properly meets the specs and can read the Unicode Table. next job for that will be to make it so that it can be given a unicode sequence and can go find it's relevent Index.

Next Job will be to implement a Console that is passed to the render context, it will need to have a proper text buffer and cursor support.

Stretch goal is to figure out a way to debug MSVC pdbs through Qemu and gdb, which may be difficult. if i can get this working, then i'll be able to properly support debugging the code, the long term goal would be support for stepping through the code directly in Visual Studios while watching in qemu.